kisanalert/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
xgboost>=1.7.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
requests>=2.28.0
python-dotenv>=0.21.0
supabase>=2.0.0
joblib>=1.2.0
# Layer 3: Deep Learning LSTM
tensorflow>=2.13.0
